Allowed customising labels
The functionality is built in the productInspect.
Removed the customisation-trigger on the etiketten page. There is no
need to program this 2 times.
Adjusted ajax call to queue a product so its compatible with the
productInspect queue-trigger.

At this moment we will not implement the 'loading of a standard text'
in the inspect plugin. as this is all based on old info. We want people
to use new stuff! This option should be removed from the inspect
interface itself.